+++ mapbox_maps_flutter_mobile/ios/mapbox_maps_flutter_mobile/Sources/mapbox_maps_flutter_mobile/Classes/Extensions.swift
@@ -465,6 +465,16 @@ extension MapboxMaps.CanonicalTileID {
     }
 }
 
+extension MapboxMaps.OverscaledTileID {
+    func toFLTOverscaledTileID() -> OverscaledTileID {
+        return OverscaledTileID(
+            overscaledZ: Int64(overscaledZ),
+            wrap: Int64(wrap),
+            canonical: canonical.toFLTCanonicalTileID()
+        )
+    }
+}
+
 extension MapboxMaps.CameraState {
     func toFLTCameraState() -> CameraState {
         return CameraState(